data/requirements.txt
# Advanced Solution Requirements
# Core dependencies
pyarrow>=10.0.0
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.19.0

